'INT' REPLACED WITH 'NUMBER'

+++ Create views for a particular table.sql
@@ -1,4 +1,4 @@
-CREATE TABLE college6161(id INT(10), name VARCHAR(20), age INT(10), address VARCHAR(10), salary INT(10));
+CREATE TABLE college6161(id NUMBER(10), name VARCHAR(20), age NUMBER(10), address VARCHAR(10), salary NUMBER(10));
 
 CREATE VIEW Gouri_view AS
 SELECT name, age
